### 2023-11-01 #6007 @editorialbot
[PRE REVIEW]: linkinglines: Using the Hough Transform to Cluster Line Segments and for Mesoscale Feature Extraction
**Submitting author:** @aikubo (<a href="http://orcid.org/0000-0002-1378-361X">Allison Kubo Hutchison</a>)
**Repository:** https://github.com/aikubo/LinkingLines
**Branch with paper.md** (empty if default branch):
**Version:** v.2.1.0
