### 2023-11-01 #6007 @kthyng
Hi @aikubo and thanks for your submission! I saw the note you submitted with the paper — was your code reviewed as part of the review process for this previously-published paper: https://agupubs.onlinelibrary.wiley.com/doi/10.1029/2022GC010842?
