### 2023-12-22 aikubo/LinkingLines#17 @evetion
Clean up repo with a .gitignore
Part of https://github.com/openjournals/joss-reviews/issues/6147

The repository has files and folders that probably don't need to be checked into version control. A non-exhaustive list:

```bash
.virtual_documents
dist
__pycache__ (several folders)
src/linkinglines.egg-info
docs/.ipynb_checkpoints
docs/.virtual_documents
docs/_build
```

You could remove these. In the future you can ignore these files with a gitignore. See:
https://git-scm.com/docs/gitignore
https://docs.github.com/en/get-started/getting-started-with-git/ignoring-files
