### 2023-12-22 aikubo/LinkingLines#18 @evetion
Documentation links don't work and examples are not formatted
Part of https://github.com/openjournals/joss-reviews/issues/6147

The documentation contains unresolved links like so: Follow this indepth [`tutorial <>`_](https://linkinglines.readthedocs.io/en/latest/index.html#id1) to get started!

If you build the documentation locally, you probably should see warnings with more details to fix these.

Furthermore, the API documentation can use some more formatting. For example in
https://linkinglines.readthedocs.io/en/latest/linkinglines.html#module-linkinglines.SyntheticLines there are multiple unformatted multiline examples that are put on a single line.
