### 2023-12-22 aikubo/LinkingLines#19 @evetion
Documentation example requires local modules
Part of https://github.com/openjournals/joss-reviews/issues/6147

```python
import sys
sys.path.append("../src")
sys.path.append("/home/akh/myprojects/Dikes_Linking_Project/dikedata/spanish peaks")

# Packages written for the paper
from htMOD import HoughTransform
from clusterMod import AggCluster
from plotmod import DotsLines, plotScatterHist, pltRec, DotsHT
from PrePostProcess import *
from examineMod import examineClusters, checkoutCluster
from fitRadialCenters import RadialFit, NearCenters, CenterFunc
```

should probably become something like

```python
import linkinglines as ll

from ll.htMOD import HoughTransform
...  # etc
```

I believe a similar thing needs to be done for your tests to work automatically.
