### 2023-12-22 aikubo/LinkingLines#20 @evetion
More generic API
Part of https://github.com/openjournals/joss-reviews/issues/6147

The codebase is currently quite fixed towards a specific research workflow. It would help re-use if it becomes more generic. Two good examples:

### On the use of WKT
Currently the package only reads and writes csv files with a (undocumented) column containing WellKnownText LineStrings. A generic version would take in any geospatial dataframe, only needing to check whether the geometries are lines. The great package [geopandas](https://geopandas.org/en/stable/), fits this use case perfectly, and automatically gives you a range of input and output files, and probably more generic methods to read coordinates from geometry.

### On the use of scripting
Ideally package code is object oriented, with methods being useful on their own. Here you use DataFrames as intermediate objects, but they still require scripting steps. This is easier to explain in an example:

Current code:
```python
data=pd.read_csv('path/to/data')
theta,rho,xc,yc=ll.HoughTransform(data)
data['theta']=theta
data['rho']=rho

dtheta=2 #degrees <-- unused
drho=500 #meters <-- unused

dikeset, Z=ll.AggCluster(data)
```

The above code requires setting `data['theta']`, before `AggCluster` can work. As such, it would be easier to let `ll.HoughTransform(data)` return a new Dataframe with these columns added, that can be used directly, like so:

```python
data = pd.read_csv('path/to/data')
ndata = ll.HoughTransform(data)  # returns dataframe with columns theta, rho, xc, yc added
dikeset, Z=ll.AggCluster(ndata)
```

You can go even one step further, and let `AggCluster` call `HoughTransform` if the required columns theta, rho etc are missing from it.
