### 2023-12-22 #6147 @evetion
Thanks for the submission, it's an interesting read and package! I have completed my review, and in my opinion it requires major revisions.

My feedback follows below, per item that I can't yet check. I've also made several suggestions in issues on the repository.

Let me know if you have questions.

## General checks
### Contribution and authorship
The code seems to have a sole author, as is stated in the `pyproject.toml` and Github. Can you explain what the roles of your co-authors in the paper are? Also, it seems to me that your documentation is partially written by a LLM, as indicated by https://github.com/aikubo/LinkingLines/blob/master/src/linkinglines/FitRadialCenters.py#L8. There's no rule against the use of LLMs, but it might be good to note its use somewhere in your documentation.

### Substantial scholarly effort
I'm in doubt whether this is a substantial scholarly effort. LoC (2825) is a bad estimator, I'm mostly looking at the possible re-use. The code has been developed for specific (published 🎉) research, but is not generalized. I've made various suggestions in https://github.com/aikubo/LinkingLines/issues/20 to improve it.

## Documentation
### Functionality documentation
It seems all the core methods are documented, but as described in https://github.com/aikubo/LinkingLines/issues/18, there are linking errors and the examples don't render nicely.

### Automated tests
There are tests 👍🏻, but AFAIK these are not automated and there are otherwise no instructions on running them.

### Community guidelines
There are no clear guidelines for third parties wishing to contribute. The use of Github issues for issues, problems, and support is implicit.

## Paper
### Summary
The summary is great, my only remark is to introduce your research domain, as `giant dike swarms` is not understood by a diverse non-specialist audience.

### A statement of need
While the paper contains a statement of need section, it mostly lists generic features, not a precise description of the gap it's trying to fill. For example, it states “simplify extraction”, but the code in question doesn't extract them.

It also does not address who the target audience is, or its relation to other work.

### State of the field
The paper does not address the current state of the field.

### Quality of writing
The writing is fine, but could use a careful reread, as some sentences are incomplete, such as l10 (sentence should probably be broken up) and l130 (`cluster fall`).
Furthermore, like the summary, large parts assume specific domain knowledge. For example, l51-63 mostly contain results from another paper. Ideally this part is rewritten (not removed), to reflect the original use of the code and its impact, but without the domain details.

### References
According to the bot post above, there are missing dois in your paper. There's also a missing key (search `survey?`).
