### 2023-12-22 aikubo/linking-and-clustering-dikes#17 @evetion
A suggestion on style (not at all required for review, but might be nice and good to know).

The current codebase is inconsistent in styling. That is, class/function names are sometimes capatalised, as are the filenames, but not always. The indentation, especially of docstrings is inconsistent, as is the whitespacing after a comma.

In Python we do suggest some [defaults](https://peps.python.org/pep-0008/), filenames are lowercased (so imports are also lowercased), as are all variables and methods, with only class names CamelCased. The practice is to use an automatic formatter that fixes the indentation and whitespacing for you. A good example is [ruff](https://docs.astral.sh/ruff/).
