### 2023-12-26 aikubo/LinkingLines#21 @nialov
Document the version of LinkingLines sent for review with tag
Part of <https://github.com/openjournals/joss-reviews/issues/6147>

You have stated that version `v.2.1.0` of `LinkingLines` is to be reviewed. You have created a GitHub release accordingly. However, there is no `git` tag representing this revision of the `LinkingLines` repository. I would suggest creating a `git` tag that represents the correct revision. If the latest commit on master represents the version to be reviewed, you can tag it with `git tag` and push the tag with `git push --tags`.
