### 2023-12-28 aikubo/LinkingLines#22 @nialov
Document development dependencies in pyproject.toml
I would document the development dependencies i.e. `pytest` in `pyproject.toml` according to the specification (<https://packaging.python.org/en/latest/specifications/pyproject-toml/#dependencies-optional-dependencies>).
