### 2023-12-28 aikubo/linking-and-clustering-dikes#22 @nialov
I would also try to follow the specification in terms of e.g. the version string in `pyproject.toml`so that a local (development) install of the package is possible.
