### 2023-12-28 aikubo/LinkingLines#25 @nialov
Add example with data other than dykes (e.g. fracture trace data)
Just to preface: use of this specific data is just a suggestion, feel free to use any other source! I do believe, however, that adding some (small) example(s) with data other than dyke data is needed to make evident the general usability of the software. No complex analysis is required for this purpose, just show how it is loaded (in `csv` form or with `geopandas`) and a simple analysis with e.g. the Cartesian space - Hough Transform space plot.

## My data suggestions

I have personally been involved with gathering fracture trace data and some of our data is publicly available:

-  <https://zenodo.org/records/7077846>

The data consists of *ESRI Shapefile* data that can be loaded with `geopandas` (See #20) or with *QGIS* and transformed to `csv` to work with the current API.

-  <https://zenodo.org/records/7919843>

This dataset is contains code and data. Fracture trace data is in `data/trace_data/traces/20m/` folder in `geojson` data type.

You can freely add individual files/parts of files to this repository as the datasets are openly and freely licensed. Just add a mention and link to *Zenodo* or the *DOI* found on the *Zenodo* pages in e.g. the `README.md` or some other suitable place.
