### 2023-12-28 #6147 @nialov
Thank you for the submission. I believe the functionality is very interesting from a geological perspective and can definitely see it being used also outside the main documented use case of dyke swarm clustering for example for examining bedrock fracture data. Therefore publishing the software in *JOSS* to allow reuse of the methodology is great!

However, it is difficult to estimate the software functionality as neither the tests or documentation (notebook) run correctly due to import issues. This is a major issue that needs to be fixed for me to assess the suitability of the package for *JOSS*. A **major revision** is therefore required.

Further issues are outlined below based on the checklist.

# General checks

## Contribution and authorship

The contributions by authors other than the main author are not explicitly stated in the paper and they can not be implicitly seen contributing to the repository on GitHub. I would like a short explanation of the roles so that I know that the contribution is not purely financial or organisational, in which case the authors should not be included.

See: <https://joss.readthedocs.io/en/latest/review_criteria.html#authorship>

## Substantial scholarly effort

I believe the scholarly effort in terms of geological method development is sufficient but this effort has already been (partly) published as part of <https://doi.org/10.1029/2022GC010842>.

As mentioned by evetion, the effort in terms of software is lacking as the described workflow is too specialized to only with with `csv` files with a strict schema. I agree with the proposed changes by evetion here: <https://github.com/aikubo/LinkingLines/issues/20> and especially recommend testing `geopandas` for reading the input files into GeoDataFrames.

## Data sharing

Data is shared but not used in the tests or documentation (notebook).

# Functionality

## Functionality

I can not confirm the functionality with the tests and notebook because they do not work.

# Documentation

## Installation instructions

I would document the development dependencies i.e. `pytest` in `pyproject.toml`.

<https://github.com/aikubo/LinkingLines/issues/22>

## Example usage

The example/recreation of the dyke analysis is great! I would however, to generalise the software, include at least some example with other data, such as with fractures.

<https://github.com/aikubo/LinkingLines/issues/25>

## Functionality documentation

All functions are documented with inputs, outputs and examples given.

However, I would expand the *Quick Start* section in the `README.md` a bit and make sure the included notebook works and is clearly linked to in documentation that users will find it.

## Automated tests

Tests are not automated or documented properly.

## Community guidelines

I suggest adding a `CONTRIBUTING.md` to let potential contributors to know if e.g. pull requests are accepted, what kind of issues should be posted, etc. Does not have to be long, but having no information on future development activity is discouraging to potential contributors.

See e.g. <https://github.com/nialov/fractopo/blob/master/CONTRIBUTING.rst>

# Software paper

The *Example Code Usage* section is not needed, code examples belong in the repository documentation, not the *JOSS* manuscript.

The *Code Structure* section is overly specific. The first paragraph is fine. Hough transform can be introduced but equations are not required for the article as they do not seem to be novel to this software/methodology. Citations to established papers with the equations and concepts are better. Details such as column names of `pandas` `DataFrames` are overly specific and need to be removed.

## Summary

The summary makes it seem like the software is readily usable for a wide variety of domains. I would suggest focusing the paper on the target audience of geoscientists for whom the software is designed.

As the summary is mainly for non-specialist audience, I would explain or give examples of geoscientific concepts such as *Linear feature analysis* and *unique feature extraction methods* immediately after introducing them when possible.

I would remove/revise the sentence on lines 12 to 15 as it is too general to be useful for the reader.

## A statement of need

I would revise the statement of need section to start with the specific need in the geological domain in relation to cluster analysis of dykes (L50-L63). Then I would at the end of the section introduce/list the other possible use cases.

I find the current lines 20-49 to be too unspecific and general and do not add any knowledge of value to the user in regards to this specific software. Parts of these lines (L20-49) can be used at the end of the section, if wanted, but otherwise I would remove most of the text in this line range.

## State of the field

The geological context is introduced but I would suggest adding context on *GIS* tools that might already conduct similar analysis if they exist. If similar software is not found I would at least show some similar packages to run specific *GIS* analyses (with Python) on (geological) data to better to show how *linkinglines* fits in the geoscientific field.

See e.g. `NetworkGT` <https://github.com/BjornNyberg/NetworkGT> or `fractopo` <https://github.com/nialov/fractopo>.

## Quality of writing

Some parts are overly general and the paper could be more concise and focused on the specific issues that the software is trying to solve.

## References

There are missing references and links, see review by evetion above.

# Further comments

Current issues and pull requests at the time of this review:

-  <https://github.com/aikubo/LinkingLines/issues/21>
-  <https://github.com/aikubo/LinkingLines/issues/22>
-  <https://github.com/aikubo/LinkingLines/issues/25>
-  <https://github.com/aikubo/LinkingLines/pull/23>
-  <https://github.com/aikubo/LinkingLines/pull/24>

Pull requests are just suggested changes, feel free to improve, comment or implement yourself!
