### 2023-12-31 #6147 @aikubo
Thank you @evetion and @nialov for the reviews!
I agree with and appreciate your comments. I will get to work generalizing the code.

On the subject of authorship, my currently listed coauthers helped in the conceptualization and theory of the code, did a little writing but I ended up rewriting most of it and prepared this paper based on our published results paper.
