+++ src/linkinglines.egg-info/SOURCES.txt
@@ -1,22 +1,0 @@
-AUTHORS
-LICENSE
-MANIFEST.in
-README.md
-pyproject.toml
-src/linkinglines/ClusterLines.py
-src/linkinglines/DilationCalculations.py
-src/linkinglines/ExamineClusters.py
-src/linkinglines/FitRadialCenters.py
-src/linkinglines/FitRectangle.py
-src/linkinglines/HT.py
-src/linkinglines/PlotUtils.py
-src/linkinglines/PrePostProcess.py
-src/linkinglines/SyntheticLines.py
-src/linkinglines/__init__.py
-src/linkinglines/__main__.py
-src/linkinglines.egg-info/PKG-INFO
-src/linkinglines.egg-info/SOURCES.txt
-src/linkinglines.egg-info/dependency_links.txt
-src/linkinglines.egg-info/requires.txt
-src/linkinglines.egg-info/top_level.txt
-tests/testClustering.py
