### 2024-03-01 b01da21
Changed HT.HoughTransform to return dataframe and updated example code in paper for issue #20
