+++ requirements.txt
@@ -1,15 +1,139 @@
+alabaster==0.7.16
+anyio==4.3.0
+argon2-cffi==23.1.0
+argon2-cffi-bindings==21.2.0
+arrow==1.3.0
+asttokens==2.4.1
+async-lru==2.0.4
+attrs==23.2.0
+Babel==2.14.0
+beautifulsoup4==4.12.3
+bleach==6.1.0
+certifi==2024.2.2
+cffi==1.16.0
+charset-normalizer==3.3.2
+click==8.1.7
+click-plugins==1.1.1
+cligj==0.7.2
+comm==0.2.1
+contourpy==1.2.0
+cycler==0.12.1
+debugpy==1.8.1
+decorator==5.1.1
+defusedxml==0.7.1
+docutils==0.19
+exceptiongroup==1.2.0
+executing==2.0.1
+fastjsonschema==2.19.1
+fiona==1.9.5
+fonttools==4.49.0
+fqdn==1.5.1
+geopandas==0.14.3
+greenlet==3.0.3
+idna==3.6
+imagesize==1.4.1
+importlib-metadata==7.0.1
+ipykernel==6.29.3
+ipython==8.22.1
+isoduration==20.11.0
+jedi==0.19.1
+Jinja2==3.1.3
+joblib==1.3.2
+json5==0.9.18
+jsonpointer==2.4
+jsonschema==4.21.1
+jsonschema-specifications==2023.12.1
+jupyter-cache==0.6.1
+jupyter-events==0.9.0
+jupyter-lsp==2.2.3
+jupyter_client==8.6.0
+jupyter_core==5.7.1
+jupyter_server==2.12.5
+jupyter_server_terminals==0.5.2
+jupyterlab==4.0.6
+jupyterlab_pygments==0.3.0
+jupyterlab_server==2.25.3
+kiwisolver==1.4.5
+markdown-it-py==2.2.0
+MarkupSafe==2.1.5
 matplotlib==3.7.2
-matplotlib_label_lines==0.6.0
+matplotlib-inline==0.1.6
+matplotlib-label-lines==0.6.0
+mdit-py-plugins==0.3.5
+mdurl==0.1.2
+mistune==3.0.2
+more-itertools==10.2.0
+myst-nb==0.17.2
+myst-parser==0.18.1
+nbclient==0.7.4
+nbconvert==7.16.1
+nbformat==5.9.2
+nbsphinx==0.9.3
+nest-asyncio==1.6.0
+notebook==7.0.4
+notebook_shim==0.2.4
 numpy==1.25.2
+overrides==7.7.0
+packaging==23.2
 pandas==2.0.3
+pandocfilters==1.5.1
+parso==0.8.3
+patsy==0.5.6
+pexpect==4.9.0
+pillow==10.2.0
+platformdirs==4.2.0
+prometheus_client==0.20.0
+prompt-toolkit==3.0.43
+psutil==5.9.8
+ptyprocess==0.7.0
+pure-eval==0.2.2
+pycparser==2.21
+Pygments==2.17.2
+pyparsing==3.0.9
 pyproj==3.6.0
-scikit_learn==1.3.0
+python-dateutil==2.9.0.post0
+python-json-logger==2.0.7
+pytz==2024.1
+PyYAML==6.0.1
+pyzmq==25.1.2
+referencing==0.33.0
+requests==2.31.0
+rfc3339-validator==0.1.4
+rfc3986-validator==0.1.1
+rpds-py==0.18.0
+scikit-learn==1.3.0
 scipy==1.11.2
 seaborn==0.12.2
+Send2Trash==1.8.2
+shapely==2.0.3
+six==1.16.0
+sniffio==1.3.1
[…]
