### 2024-03-05 7747a22
add dev dependencies pyproject #22
