+++ tests/testClustering.py
@@ -6,6 +6,11 @@
 # @author: akh
 # """
 
+import numpy as np
+import pandas as pd
+import linkinglines as ll
+import pytest
+
 def testValidity(lines, dtheta, drho):
     
     if any(lines['ThetaRange'] > dtheta):
@@ -18,446 +23,60 @@ def testValidity(lines, dtheta, drho):
     else: 
         print("Passed Theta Validity Check 1")
 
-def Run3Times(dikeset, dtheta, drho, plot=False, **kwargs):
-    xc1,yc1=HT_center(dikeset)
+def Run3Times(dikeset, dtheta, drho, **kwargs):
+    xc1,yc1=ll.HT_center(dikeset)
     # run the first time 
-    dikeset, Z1=HT_AGG_custom(dikeset, dtheta, drho, **kwargs)
-    lines1, IC1=examineClusters(dikeset)
+    dikeset, Z1=ll.AggCluster(dikeset, dtheta, drho, **kwargs)
+    lines1, IC1=ll.examineClusters(dikeset)
     
     # rotate 45 deg 
-    dikeset2=rotateData2(dikeset, 45)
-    theta2, rho2, xc2, yc2=AKH_HT(dikeset2)
-    dikeset2['theta']=theta2
-    dikeset2['rho']=rho2
-    dikeset2, Z2=HT_AGG_custom(dikeset2, dtheta, drho, **kwargs)
-    dikeset2=rotateData2(dikeset2,-45)
-    lines2, IC2=examineClusters(dikeset2)
+    dikeset2=ll.rotateData(dikeset, 45)
+    dikeset2, xc2, yc2=ll.HT(dikeset2)
+    dikeset2, Z2=ll.AggCluster(dikeset2, dtheta, drho, **kwargs)
+    dikeset2=ll.rotateData(dikeset2,-45)
+    lines2, __=ll.examineClusters(dikeset2)
     
     #move to lower left
     dx=np.min([dikeset['Xstart'].min(), dikeset['Xend'].min()])
     dy=np.min([dikeset['Ystart'].min(), dikeset['Yend'].min()])
-    theta3, rho3, xc3, yc3=AKH_HT(dikeset, xc=dx,yc=dy)
-    dikeset3=dikeset.copy()
-    dikeset3['theta']=theta3
-    dikeset3['rho']=rho3
+    dikeset3, xc3, yc3=ll.HT(dikeset, xc=dx,yc=dy)
+
+    dikeset3, Z3=ll.AggCluster(dikeset3, dtheta, drho, **kwargs)
     
-    dikeset3, Z3=HT_AGG_custom(dikeset3, dtheta, drho, **kwargs)
-    
-    lines3, IC3=examineClusters(dikeset3)
+    lines3, IC3=ll.examineClusters(dikeset3)
     
     #check the changes 
-    eq12, diff12=checkClusterChange(lines1, lines2)
-    eq13, diff13=checkClusterChange(lines1, lines3)
-    eq32, diff32=checkClusterChange(lines3, lines2)
+    eq12, diff12=ll.checkClusterChange(lines1, lines2)
+    eq13, diff13=ll.checkClusterChange(lines1, lines3)
+    eq32, diff32=ll.checkClusterChange(lines3, lines2)
     
-    Flines1=FilterLines(lines1)
-    Flines2=FilterLines(lines2)
-    Flines3=FilterLines(lines3)
+    Flines1=ll.FilterLines(lines1)
+    Flines2=ll.FilterLines(lines2)
+    Flines3=ll.FilterLines(lines3)
     
     print("Comparing the filtered lines")
-    eq12, diff12=checkClusterChange(Flines1, Flines2)
[…]
