### 2024-03-05 f3fc646
update readme with new api #20
