### 2024-03-05 a73435d
Added contributing.md for openjournals/joss-reviews/#6147
