### 2024-03-05 5a66a04
update tests with new api, resolves #20
