+++ tests/preprocessing_test.py
@@ -8,13 +8,13 @@ Created on Sat Sep 9 10:01:18 2023
 
 # Import necessary modules and libraries
 from linkinglines import HoughTransform, CyclicAngleDist, rotateData, HT_center
-from linkinglines import fromHT, fragmentDikes, WKTtoArray, readFile
+from linkinglines import fromHT, fragmentDikes, WKTtoArray, readFile, writeFile
 
 import pytest
 import pandas as pd
 import geopandas as gpd
 import numpy as np
-
+import os
 
 class TestHTCenter:
     # Test case 1: Test with a DataFrame containing horizontal line segments
@@ -64,23 +64,23 @@ class TestHTCenter:
 class TestHough:
     # Test case for an empty DataFrame input to HoughTransform function
     @staticmethod
-    def test_emptyDataFrame_akhht():
+    def test_emptyDataFrame_ht():
         # Create an empty DataFrame with the required columns
         empty_df = pd.DataFrame(columns=['Xstart', 'Ystart', 'Xend', 'Yend'])
         
         # Call the HoughTransform function with the empty DataFrame
         with pytest.raises(ValueError):
-            theta, rho, xc, yc = HoughTransform(empty_df)
+            df, xc, yc = HoughTransform(empty_df)
     
     # Test case for a non-DataFrame input to HoughTransform function
     @staticmethod
-    def test_notDataFrame_akhht():
+    def test_notDataFrame_ht():
         # Create an array (non-DataFrame)
         array = np.linspace(0, 100)
         
         # Call the HoughTransform function with the non-DataFrame input
         with pytest.raises(ValueError):
-            theta, rho, xc, yc = HoughTransform(array)
+            df, xc, yc = HoughTransform(array)
             
     # Test case for points not lines
     @staticmethod
@@ -94,7 +94,7 @@ class TestHough:
 
         # Call the HoughTransform function with the points input
         with pytest.raises(ValueError):
-            theta, rho, xc, yc = HoughTransform(data)
+            data, xc, yc = HoughTransform(data)
 
     # Test case for HoughTransform with specified center coordinates
     @staticmethod
@@ -111,7 +111,7 @@ class TestHough:
         xc, yc = 1.5, 1.5
         
         # Calculate the Hough Transform
-        theta, rho, result_xc, result_yc = HoughTransform(data, xc=xc, yc=yc)
+        df, result_xc, result_yc = HoughTransform(data, xc=xc, yc=yc)
         
         # Perform assertions
         assert result_xc == xc
@@ -120,9 +120,11 @@ class TestHough:
         # Check calculated theta and rho values
         theta_true = np.array([-45., -45., -45., -45.])
         rho_true = np.array([0, 0, 0, 0])
-        assert len(theta) == len(data)
+
+        theta = df['theta'].values
+        rho = df['rho'].values
+        assert len(df)==len(data)
         assert all([np.isclose(a, b) for a, b in zip(theta, theta_true)])
-        assert len(rho) == len(data)
         assert all([np.isclose(a, b) for a, b in zip(rho, rho_true)])
     
     # Test case for HoughTransform without specifying center coordinates
@@ -137,16 +139,16 @@ class TestHough:
         })
         
         # Calculate the Hough Transform without specifying center
-        theta, rho, xc, yc = HoughTransform(data)
+        df, xc, yc = HoughTransform(data)
         
         # Perform assertions
-        assert isinstance(theta, np.ndarray)
-        assert isinstance(rho, np.ndarray)
[…]
