### 2024-03-14 4c948d4
add contributing to read me, remove md file
