+++ src/linkinglines/FitRadialCenters.py
@@ -5,33 +5,21 @@
 # -*- coding: utf-8 -*-
 
 '''
-The provided Python module appears to be a collection of functions for analyzing
- and processing dike lines data in a geospatial context. Here's a summary of the
- module's main functionalities and the purpose of each function:
+FitRadialCenters Module
 
+This module provides functions for fitting a radial model to dike lines data, calculating radial azimuthal angles, measuring the "clumpiness" of radial dike swarm angles, and more.
+
+Functions:
 - `CenterFunc(theta, xr, yr, xc, yc)`: Calculates the radial distance of a point from a given center for a specified angle.
-
 - `RadialFit(lines, plot=False, ColorBy=None, weight='LayerNumber', ThetaRange=[-90, 90], xc=None, yc=None)`: Fits a radial model to dike lines data, optionally visualizing the fit. It returns center information such as coordinates and goodness of fit.
-
 - `RadialAzimuthal(lines, Center)`: Calculates radial azimuthal angles of dike lines relative to a specified center.
-
 - `CyclicAngle360(u, v)`: Computes the cyclic angle between two angles in the range [0, 360).
-
 - `AngleSpacing(rAngle)`: Calculates statistics related to the spacing between radial azimuthal angles.
-
 - `RipleyRadial(rAngle, plot=False)`: Measures the "clumpiness" of radial dike swarm angles using the Ripley K function, with an option to visualize the results.
-
 - `ExpandingR(lines, Center)`: Measures the density of lines at different radial distances from a center.
-
 - `NearCenters(lines, Center, tol=10000, printOn=False)`: Identifies lines near a specified center within a tolerance, providing detailed information about those lines.
-
 - `writeCenterWKT(df, name)`: Writes center information to a Well-Known Text (WKT) file, allowing for easy geospatial data export.
-Overall, this module is designed for the analysis of geological dike structures,
-particularly in terms of their orientation and distribution relative to a central
- point. It includes functions for fitting radial models, assessing clustering patterns,
- and exporting results in a geospatial format.
 
- written by akh
 '''
 import numpy as np
 import pandas as pd
@@ -40,23 +28,40 @@ from scipy.spatial.distance import pdist, squareform
 import matplotlib.pyplot as plt
 from .PrePostProcess import whichForm
 from .PlotUtils import DotsLines
+import os
+import geopandas as gpd
 
 
 def CenterFunc(theta, xr, yr, xc, yc):
     """
     Calculate the radial distance of a point from a center given an angle.
 
-    Parameters:
-        theta (float): Angle in degrees.
-        xr (float): X-coordinate of the point.
-        yr (float): Y-coordinate of the point.
-        xc (float): X-coordinate of the center.
-        yc (float): Y-coordinate of the center.
+    This function computes the radial distance of a point (or points) from a specified center based on
+    a given angle. The radial distance, `rhoRadial`, is determined from the cartesian coordinates of the
+    point and the center, along with the angle in degrees between them. This calculation is useful for
+    geometric analyses and transformations in a polar coordinate system.
 
-    Returns:
-        float: Radial distance of the point from the center.
+    Parameters
+    ----------
+    theta : float or array_like
+        Angle(s) in degrees between the point(s) and the horizontal axis.
+    xr : float
+        X-coordinate of the point.
+    yr : float
+        Y-coordinate of the point.
+    xc : float
+        X-coordinate of the center.
+    yc : float
+        Y-coordinate of the center.
+
+    Returns
+    -------
+    rhoRadial : float or ndarray
+        The radial distance(s) of the point(s) from the center. If `theta` is a single float, `rhoRadial`
+        will be a single float. If `theta` is an array_like, `rhoRadial` will be an ndarray of the same shape.
 
     """
+
[…]
