+++ src/linkinglines/DilationCalculations.py
@@ -19,58 +19,45 @@ def dilation(df, binWidth=1, averageWidth=1, method='Expanded'):
     """
     Calculate dilation values for a given dataset along EW and NS directions.
 
-    This function calculates dilation values along the East-West (EW) and North-South (NS) directions based on the input DataFrame. It divides the data into bins and computes the dilation for each bin. The method parameter allows you to choose between different dilation calculation methods.
+    This function calculates dilation values along the East-West (EW) and North-South (NS) directions based on the input DataFrame. It divides the data into bins and computes the dilation for each bin.
+    The method parameter allows you to choose between different dilation calculation methods.
+
+     The differences between the three dilation calculation methods (Expanded, Average, and Total) lie in how they compute dilation values from the data. Dilation is a measure of how much a geological feature has been stretched or expanded, typically expressed as a ratio. Let's explain the differences between these methods:
+
+    1. **Expanded Dilation:**
+    - **Method Description:** Expanded dilation calculates dilation by considering how much a line segment expands within a bin. It accounts for segments that partially cross the bin's boundaries and calculates dilation accordingly.
+    - **Calculation:** For each bin, it sums the fractional expansions of line segments within the bin. Fractional expansion is the portion of a line segment that enters the bin. It considers both the portion of segments that start or end within the bin and the portions that cross into the bin.
+
+    2. **Average Dilation:**
+    - **Method Description:** Average dilation calculates dilation by averaging the dilation values of all line segments entirely contained within each bin.
+    - **Calculation:** For each bin, it calculates the dilation value for each segment entirely contained within the bin. It then averages these segment dilation values to obtain the bin's dilation value.
+
+    3. **Total Dilation:**
+    - **Method Description:** Total dilation calculates dilation by summing the dilation values of all line segments entirely contained within each bin.
+    - **Calculation:** Similar to Average Dilation, it calculates the dilation value for each segment entirely contained within the bin. Instead of averaging, it sums these segment dilation values to obtain the bin's dilation value.
+
+    In summary, the choice between Expanded, Average, or Total dilation depends on your specific geological analysis needs. Expanded dilation is more detailed and suitable for capturing partial segment contributions. Average dilation provides a smoother representation, while Total dilation gives a cumulative measure of dilation within each bin. The choice should be based on the level of detail and accuracy required for your analysis.
+    
 
     Parameters:
+    ----------
         df (pandas.DataFrame): The input DataFrame containing data.
         binWidth (float, optional): The width of bins used for dilation calculation. Default is 1.
         averageWidth (float, optional): The average width used in dilation calculation. Default is 1.
         method (str, optional): The dilation calculation method. Options are 'Average', 'Total', and 'Expanded'. Default is 'Expanded'.
 
- The differences between the three dilation calculation methods (Expanded, Average, and Total) lie in how they compute dilation values from the data. Dilation is a measure of how much a geological feature has been stretched or expanded, typically expressed as a ratio. Let's explain the differences between these methods:
-
-1. **Expanded Dilation:**
-
-   - **Method Description:** Expanded dilation calculates dilation by considering how much a line segment expands within a bin. It accounts for segments that partially cross the bin's boundaries and calculates dilation accordingly.
-
-   - **Calculation:** For each bin, it sums the fractional expansions of line segments within the bin. Fractional expansion is the portion of a line segment that enters the bin. It considers both the portion of segments that start or end within the bin and the portions that cross into the bin.
-
[…]

[…]
