+++ src/linkinglines/PrePostProcess.py
@@ -43,17 +43,23 @@ import os
 import geopandas
 
 def readFile(name, preprocess=True):
+    """
+    Reads in a file and returns a pandas dataframe.
+
+    Parameters
+    ----------
+    name : str
+        The path to the file to be read in.
+
+    preprocess : bool, optional
+        Indicates whether to preprocess the data, by default True.
+
+    Returns
+    -------
+    data : pandas.DataFrame
+        A pandas or geopandas dataframe containing the read data.
 
     """
-    Reads in a file and returns a pandas dataframe
-
-    Parameters:
-        name: (string) the path to the file to be read in
-    
-    Returns:
-        data: (pandas.DataFrame) a pandas or geopandas dataframe
-    """
-
     # if not a valid path, return error
     if not os.path.exists(name):
         raise ValueError("Invalid path")
@@ -87,15 +93,15 @@ def midPoint(df):
 
     Parameters
     ----------
-    df: pandas.Dataframe
-        dataframe of the line segments
-        must contain ["Xstart", "Ystart", "Xend", "Yend"]
+    df : pandas.DataFrame
+        Dataframe containing the line segments with columns ["Xstart", "Ystart", "Xend", "Yend"].
 
     Returns
     -------
-    df: pandas.Dataframe
-    with new columns of ['Xmid', 'Ymid']
+    df : pandas.DataFrame
+        Dataframe with new columns ['Xmid', 'Ymid'] indicating the midpoints of the line segments.
     """
+
     df['Xmid']=(df['Xstart']+df['Xend'])/2
     df['Ymid']=(df['Ystart']+df['Yend'])/2
 
@@ -104,20 +110,22 @@ def midPoint(df):
 
 def writeToWKT(df,name, myProj=None):
     """
+    Writes a dataframe to a CSV file using Well-Known Text (WKT) format for line vectors.
 
-    Writes a dataframe to a CSV file in the format of a QGIS layer.
-    Uses well known text (WKT) to write the data as readable line vectors in QGIS.
-    Uses the WKTtoArray function to process the dataframe.
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        Dataframe with columns ['Xstart', 'Ystart', 'Xend', 'Yend', 'seg_length'].
+    name : str
+        Name of the output file.
+    myProj : projection, optional
+        Projection of the dataframe. If None, WGS84 is assumed, by default None.
 
-    Parameters:
-        df: a pandas dataframe with columns Xstart,Ystart,Xend,Yend,seg_length
-        name: the name of the file to be written
-        myProj: the projection of the dataframe. If none is given, the projection is set to WGS84
-        Returns:
-            A CSV file with the dataframe in the format of a QGIS layer
-            """
-
-    #if myProj not None:
+    Returns
+    -------
+    pandas.DataFrame
+        The input dataframe with an additional 'Linestring' column.
+    """
 
     front="LINESTRING("
     linestring=[]
@@ -134,15 +142,23 @@ def writeToWKT(df,name, myProj=None):
 
 def writetoGeoData(df, name, driver, myProj=None):
     """
-    Writes a dataframe to a shapefile.
+    Writes a dataframe to a geospatial file (shapefile, geopackage, or geojson).
 
-    Parameters:
-        df: a pandas dataframe with columns Xstart,Ystart,Xend,Yend,seg_length
-        name: the name of the file to be written
-        myProj: the projection of the dataframe. If none is given, the projection is set to WGS84
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        Dataframe with geospatial data to be written.
+    name : str
+        Name of the output file.
+    driver : str
+        Format of the file (e.g., 'ESRI Shapefile', 'GeoJSON').
+    myProj : projection, optional
[…]
