+++ src/linkinglines/PlotUtils.py
@@ -37,38 +37,48 @@ from matplotlib.colors import LogNorm
 np.random.seed(5)
 
 import matplotlib.ticker as mticker
-
-class FixCartesianLabels():
+class FixCartesianLabels:
     """
-    Moves Offset axis tick label to axis label.
+    Class for adjusting axis labels by moving the offset of axis tick labels to the axis label.
 
-    This class is used to adjust the axis labels by moving the offset (scale factor) of the axis tick labels to the axis label itself. It's particularly helpful when dealing with plots where the offset notation is desired to be shown as part of the axis label.
+    This class is designed to assist in adjusting the display of axis labels in plots by incorporating the
+    offset (scale factor) of the axis tick labels into the axis label itself. This adjustment is particularly
+    useful in cases where the offset notation is preferred to be part of the axis label, enhancing the
+    readability and presentation of plots.
 
-    Attributes:
-        None
+    Parameters
+    ----------
+    ax : matplotlib.axis.Axis
+        The axis object for which the labels are to be adjusted.
 
-    Methods:
-        __init__(self, ax):
-            Initializes an instance of the FixCartesianLabels class for a given axis.
+    Methods
+    -------
+    update(ax, lim)
+        Updates the axis labels by incorporating the offset into the axis label text.
 
-        update(self, ax, lim):
-            Updates the axis labels by moving the offset to the axis label.
+    Examples
+    --------
+    >>> import matplotlib.pyplot as plt
+    >>> fig, ax = plt.subplots()
+    >>> ax.plot([1, 2, 3], [1e6, 2e6, 3e6])
+    >>> y_label_fixer = FixCartesianLabels(ax.yaxis)
+    >>> plt.show()
 
-    Example:
-        import matplotlib.pyplot as plt
-
-        # Create a sample plot
-        fig, ax = plt.subplots()
-        ax.plot([1, 2, 3], [1e6, 2e6, 3e6])
-
-        # Initialize FixCartesianLabels for the y-axis
-        y_label_fixer = FixCartesianLabels(ax.yaxis)
-
-        plt.show()
+    Note
+    ----
+    The `update` method is connected to the axis's 'Cartesian Plots Updated' event and is triggered
+    whenever the axis is updated, ensuring that the label adjustments are applied automatically.
     """
+
     def __init__(self, ax):
-        # self.axis = {"y":ax.yaxis, "x":ax.xaxis}[axis]
-        # self.labelx=""
+        """
+        Initializes the FixCartesianLabels instance for the specified axis.
+
+        Parameters
+        ----------
+        ax : matplotlib.axis.Axis
+            The axis for which the labels should be updated.
+        """
         ax.callbacks.connect('Cartesian Plots Updated', self.update)
         ax.figure.canvas.draw()
         self.update(ax, None)
@@ -77,116 +87,118 @@ class FixCartesianLabels():
         """
         Updates the axis labels by moving the offset to the axis label.
 
-        Args:
-            ax (matplotlib.axis.Axis): The axis for which the labels should be updated.
-            lim: Unused parameter (needed for the callback).
+        This method adjusts the axis labels by incorporating the offset (scale factor) from the axis tick labels
+        into the axis label itself. The update is applied to both the x and y axes.
 
-        Returns:
-            None
+        Parameters
+        ----------
+        ax : matplotlib.axis.Axis
+            The axis for which the labels are to be updated.
+        lim : unused
+            An unused parameter, included for compatibility with callback requirements.
+
+        Returns
+        -------
+        None
         """
         for i, l in zip([ax.yaxis, ax.xaxis], ['Y', 'X']):
[…]

[…]
