### 2024-03-18 babc088
added poetry as manager, all test pasing
