### 2024-03-18 8b0392a
fixed documentation formatting and autodoc
