### 2024-03-18 5cdf1e7
Refactor ExamineClusters.py to improve readability and add missing type annotations
