+++ src/linkinglines/FitRectangle.py
@@ -41,28 +41,26 @@ def rotateXYShift(ang,x,y,h,k):
   ----------
   ang : float
       The angle in radians by which to rotate the coordinates.
-
   x : float or numpy array
       The x-coordinate of the point to be transformed.
-
   y : float or numpy array
       The y-coordinate of the point to be transformed.
-
   h : float
       The x-coordinate of the center point about which the rotation and shift will be performed.
-
   k : float
       The y-coordinate of the center point about which the rotation and shift will be performed.
 
   Returns
   -------
-  float, float or numpy array, numpy array
-      The transformed coordinates (xp, yp) after rotating and shifting (x, y) about (h, k).
+  xp : float or numpy array
+    The transformed x-coordinate after rotating and shifting (x, y) about (h, k).
+  yp : float or numpy array
+    The transformed y-coordinate after rotating and shifting (x, y) about (h, k).
 
   Notes
   -----
-  - The function rotates the point (x, y) counterclockwise by 'ang' radians around the center point (h, k).
-  - It returns the new coordinates (xp, yp) after the transformation.
+  The function rotates the point (x, y) counterclockwise by 'ang' radians around the center point (h, k).
+
   """
 
     xp= (x-h)*np.cos(ang)-(y-k)*np.sin(ang)
@@ -100,24 +98,15 @@ def unrotate(ang, x, y, h, k):
     -----
     - The function reverses the rotation and shift applied to the point (x, y) by 'ang' radians around the center point (h, k).
     - It returns the original coordinates (xr, yr) before the transformation.
+
+    See Also
+    --------
+    rotateXYShift : Rotate and shift coordinates (x, y) about a center point (h, k) by an angle 'ang'.
+
     """
     xr = x * np.cos(ang) + y * np.sin(ang) + h
     yr = -1 * (x) * np.sin(ang) + y * np.cos(ang) + k
     return xr, yr
-
-
-
-def inRectangle(a,b,xp,yp):
-    xtop=b; ytop=a
-    insideX=True
-    insideY=True
-
-    if any(abs(xp) > xtop):
-        insideX=False
-    if any(abs(yp) > ytop):
-        insideY=False
-
-    return insideX, insideY
 
 
 def endpoints(lines):
@@ -131,14 +120,11 @@ def endpoints(lines):
 
     Returns
     -------
-    numpy.ndarray, numpy.ndarray
-        Two NumPy arrays containing the x and y coordinates of endpoints, respectively.
+    xlist : numpy.ndarray
+        A NumPy array containing the x coordinates of endpoints.
+    ylist : numpy.ndarray
+        A NumPy array containing the y coordinates of endpoints.
 
-    Notes
-    -----
-    - This function extracts the endpoint coordinates from the DataFrame 'lines'.
-    - The DataFrame 'lines' should have columns 'Xstart', 'Ystart', 'Xend', and 'Yend' to represent line segments.
-    - The x and y coordinates of all endpoints are stored in separate NumPy arrays and returned.
     """
     xlist = np.array([])
     ylist = np.array([])
@@ -167,14 +153,10 @@ def midpoint(lines):
 
     Returns
     -------
-    numpy.ndarray, numpy.ndarray
-        Two NumPy arrays containing the x and y coordinates of midpoints, respectively.
-
-    Notes
-    -----
-    - This function calculates the midpoint coordinates of each line segment in the DataFrame 'lines'.
-    - The midpoint of a line is computed as the average of the x and y coordinates of its endpoints.
-    - The resulting x and y coordinates are stored in separate NumPy arrays and returned.
+    xlist : numpy.ndarray
+        A NumPy array containing the x coordinates of endpoints.
+    ylist : numpy.ndarray
[…]

[…]
