+++ src/linkinglines/FitRadialCenters.py
@@ -1,6 +1,6 @@
 # LinkingLines Package
  # Written by aikubo
- # Version: 2.1.0
+ 
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
