+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "linkinglines"
-version = "2.1.0"
+version = "2.1.1"
 description = "Process linear geospatial data and link them using the Hough Transform and Agglomerative Clustering"
 authors = ["aikubo <alli.i.kubo@gmail.com>"]
 license = "MIT"
@@ -8,6 +8,7 @@ readme = "README.md"
 homepage = "https://github.com/aikubo/LinkingLines"
 repository = "https://github.com/aikubo/LinkingLines"
 documentation = "https://linkinglines.readthedocs.io/en/latest/"
+keywords = ["geospatial", "houghtransform", "clustering"]
 
 [tool.poetry.dependencies]
 python = "^3.10"
@@ -46,6 +47,12 @@ jupyterlab = "^4.1.5"
 pytest = "^8.1.1"
 pytest-datafiles = "^3.0.0"
 pytest-datadir = "^1.5.0"
+
+[tool.poetry_bumpversion.file."src/linkinglines/__init__.py"]
+[tool.poetry_bumpversion.file."src/linkinglines/__main__.py"]
+[tool.poetry_bumpversion.file."src/linkinglines/ClusterLines.py"]
+
+
 
 [build-system]
 requires = ["poetry-core"]
