+++ .github/workflows/docsRebuilt.yml
@@ -1,4 +1,4 @@
-on: [pull_request, workflow_dispatch]
+on: [push, pull_request, workflow_dispatch]
 
 jobs:
   build:
