+++ .github/workflows/docsRebuilt.yml
@@ -6,9 +6,9 @@ jobs:
     steps:
     - uses: actions/checkout@v2
     - name: Set up Python 3.10
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v5
       with:
-        python-version: 3.10
+        python-version: "3.10"
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
