### 2024-03-19 a8fca17
Update docsRebuilt.yml
update secrets handling
