### 2024-03-19 4a76419
Update conf.py with correct extension name
