+++ .github/workflows/docsRebuilt.yml
@@ -10,6 +10,10 @@ jobs:
       uses: actions/setup-python@v5
       with:
         python-version: "3.10"
+    - name: Install Pandoc
+      run: |
+        sudo apt-get update
+        sudo apt-get install -y pandoc
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
