+++ .github/workflows/docsRebuilt.yml
@@ -30,6 +30,6 @@ jobs:
     - name: Trigger RTDs build
       uses: dfm/rtds-action@v1
       with:
-        webhook_url: ${{ secrets.RTDS_WEBHOOK_URL }}
-        webhook_token: ${{ secrets.RTDS_WEBHOOK_TOKEN }}
+        webhook_url: ${{ secrets.RTD_WEBHOOK_URL }}
+        webhook_token: ${{ secrets.RTD_WEBHOOK_TOKEN }}
         commit_ref: ${{ github.ref }}
