+++ .github/workflows/AutomatedTests.yml
@@ -1,0 +1,35 @@
+name: Python Tests
+
+on:
+  push:
+    branches: [ main, JOSS ]
+  pull_request:
+    branches: [ main, JOSS ]
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python 3.10
+      uses: actions/setup-python@v5
+      with:
+        python-version: "3.10"
+
+    - name: Install Poetry
+      run: |
+        pip install --upgrade pip
+        pip install poetry
+
+    - name: Configure Poetry
+      run: |
+        poetry config virtualenvs.create false
+
+    - name: Install dependencies
+      run: |
+        poetry install
+
+    - name: Run tests with pytest
+      run: |
+        poetry run pytest
+
