### 2024-03-19 32b1a4b
Update AutomatedTests.yml
