### 2024-03-19 c60cf8e
Update AutomatedTests.yml
