+++ config.toml
@@ -1,4 +1,0 @@
-# config.toml
-
-[feed]
-url = "https://github.com/aikubo/Linking-and-Clustering-Dikes"
