### 2024-03-19 e738bea
update paper based on comments in https://github.com/openjournals/joss-reviews/issues/6147
