+++ paper/paper.md
@@ -11,17 +11,7 @@ authors:
   - name: Allison Kubo Hutchison
     orcid: 0000-0002-1378-361X
     affiliation: 1
-  - name: Leif Karlstrom
-    orcid: 0000-0002-2197-2349
-    affiliation: 1
-  - name: Tushar Mittal
-    orcid: 0000-0002-8026-0018
-    affiliation: 2
-affiliations:
- - name: Department of Earth Sciences, University of Oregon, Eugene, OR, USA
-   index: 1
- - name: Department of Geosciences, Pennsylvania State University, University Park, PA, USA
-   index: 2
+
 date: 5 October 2023
 bibliography: paper.bib
 header-includes:
@@ -30,47 +20,35 @@ header-includes:
 
 # Summary
 
-Linear feature analysis plays a fundamental role in various scientific and geospatial applications, from detecting infrastructure networks to characterizing geological formations. In this paper, we introduce `linkinglines`, an open-source Python package tailored for the clustering, and feature extraction of linear structures in geospatial data. Our package leverages the Hough Transform, commonly used in image processing, performs clustering of line segments in the Hough Space then provides unique feature extraction methods and visualization. `linkinglines` empowers researchers, data scientists, and analysts across diverse domains to efficiently process, understand, and extract valuable insights from linear features, contributing to more informed decision-making and enhanced data-driven exploration. We have used `linkinglines` to map giant dike swarms in @kubo2023. This JOSS paper provides an in-depth overview of the package's design, functionalities, and practical use cases, highlighting its importance as a versatile tool in geospatial data analysis.
+Linear feature analysis plays a fundamental role in various scientific and geospatial applications, from detecting infrastructure networks to characterizing geological formations. In this paper, we introduce `linkinglines`, an open-source Python package tailored for the clustering, and feature extraction of linear structures in geospatial data. Our package leverages the Hough Transform, commonly used in image processing, performs clustering of line segments in the Hough Space then provides unique feature extraction methods and visualization. `linkinglines` empowers researchers, data scientists, and analysts across diverse domains to efficiently process, understand, and extract valuable insights from linear features, contributing to more informed decision-making and enhanced data-driven exploration. We have used `linkinglines` to map dike swarms with thousands of segments associated with Large Igneous Provinces in @kubo2023. T
 
 # Statement of Need
 
-The `linkinglines` Python package addresses the need for efficient and accurate line clustering and analysis in geospatial and image data processing in addition to adding feature extraction capabilities. As the volume of data continues to grow across various domains, including remote sensing, computer vision, and geographic information systems (GIS), there is an increasing demand for tools that can simplify the extraction and analysis of linear features such as dikes, fractures, roads, rivers, and infrastructure.
+The `linkinglines` Python package addresses the need for quantitative and automated line clustering and analysis in geologic data. As the volume of data continues to grow, including remote sensing, computer vision, and geographic information systems (GIS), there is an increasing demand for tools that can simplify the extraction and analysis of linear features such as dikes, fractures, roads, lineaments.
 
 The primary needs that the `linkinglines` package fulfills include:
 
-1. **Dissected Line Extraction or Data Reduction**: In areas where land cover or data availabilty effects the complete mapping of linear features, this package can link together similarly oriented segments into lines. This is also an effective data reduction technique. The `linkinglines` package offers algorithms and utilities for identifying and extracting such features from complex data.
+1. **Dissected Line Extraction or Data Reduction**: In areas where land cover or data availabilty effects the complete mapping of linear features, this package can link together similarly oriented segments into lines in a quantitiative automated way. This is also an effective data reduction technique.
 
[…]
