+++ paper/paper.bib
@@ -1,17 +1,13 @@
-@misc{iso2016information,
-  title={Information Technology--Database Languages--SQL Multimedia and Application Packages--Part 3: Spatial},
-  author={ISO/IEC 13249-3},
-  year={2016},
-  publisher={International Organization for Standardization, International~…}
-}
-@article{hough1962method,
-  title={A method and means for recognition of complex patterns},
-  author={Hough, Paul V. C.},
-  journal={US Patent},
-  year={1962},
-  volume={3},
-  pages={465-467},
-  publisher={Google Patents}
+@article{hough,
+  title={Use of the Hough transformation to detect lines and curves in pictures},
+  author={Duda, Richard O and Hart, Peter E},
+  journal={Communications of the ACM},
+  volume={15},
+  number={1},
+  pages={11--15},
+  year={1972},
+  publisher={ACM New York, NY, USA}
+  doi ={10.1145/361237.361242}
 }
 @article{ballard1981generalizing,
   title={Generalizing the Hough transform to detect arbitrary shapes},
@@ -22,25 +18,36 @@
   pages={111-122},
   year={1981},
   publisher={Elsevier}
+  doi ={10.1016/0031-3203(81)90009-1}
 }
-@article{pandas,
-  title={Data structures for statistical computing in python},
-  author={McKinney, Wes},
-  journal={Proceedings of the 9th Python in Science Conference},
-  volume={445},
-  pages={51--56},
-  year={2010},
-  organization={Austin, TX}
+@inproceedings{pandas,
+ author = {McKinney, Wes},
+ title = {Data Structures for Statistical Computing in Python},
+ pages = {56--61},
+ bookpagination = {page},
+ publisher = {SciPy},
+ series = {Proceedings of the Python in Science Conference},
+ editor = {{van der Walt}, St{\'e}fan and Millman, Jarrod},
+ booktitle = {Proceedings of the 9th Python in Science Conference},
+ year = {2010},
+ doi = {10.25080/Majora-92bf1922-00a},
+ booksubtitle = {SciPy 2010},
+ eventtitle = {Python in Science Conference},
+ venue = {Austin, Texas},
+ eventdate = {June 28 - July 3 2010}
 }
-@article{scipy,
-  title={Scipy: open source scientific tools for python},
-  author={Jones, Eric and Oliphant, Travis and Peterson, Pearu and others},
-  journal={Computational Science \& Discovery},
-  volume={2},
-  number={1},
-  pages={22},
-  year={2011},
-  publisher={IOP Publishing}
+
+
+@article{virtanen2020scipy,
+  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
+  journal={Nature methods},
+  volume={17},
+  number={3},
+  pages={261--272},
+  year={2020},
+  publisher={Nature Publishing Group}
+  doi ={10.1038/s41592-019-0686-2}
 }
 @article{kubo2023,
   title={Multiscale spatial patterns in giant dike swarms identified through objective feature extraction},
@@ -51,6 +58,7 @@
   pages={e2022GC010842},
   year={2023},
   publisher={Wiley Online Library}
+  doi ={10.1029/2022GC010842}
 }
 @article{morriss2020,
   title={The Chief Joseph dike swarm of the Columbia River flood basalts, and the legacy data set of William H. Taubeneck},
@@ -61,4 +69,141 @@
   pages={1082--1106},
   year={2020},
   publisher={Geological Society of America}
[…]
