+++ paper/paper.bib
@@ -6,8 +6,8 @@
   number={1},
   pages={11--15},
   year={1972},
-  publisher={ACM New York, NY, USA}
-  doi ={10.1145/361237.361242}
+  publisher={ACM New York, NY, USA},
+  doi = {10.1145/361237.361242}
 }
 @article{ballard1981generalizing,
   title={Generalizing the Hough transform to detect arbitrary shapes},
@@ -17,8 +17,8 @@
   number={2},
   pages={111-122},
   year={1981},
-  publisher={Elsevier}
-  doi ={10.1016/0031-3203(81)90009-1}
+  publisher={Elsevier},
+  doi = {10.1016/0031-3203(81)90009-1}
 }
 @inproceedings{pandas,
  author = {McKinney, Wes},
@@ -46,7 +46,7 @@
   number={3},
   pages={261--272},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
   doi ={10.1038/s41592-019-0686-2}
 }
 @article{kubo2023,
@@ -57,7 +57,7 @@
   number={9},
   pages={e2022GC010842},
   year={2023},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi ={10.1029/2022GC010842}
 }
 @article{morriss2020,
@@ -68,7 +68,7 @@
   number={4},
   pages={1082--1106},
   year={2020},
-  publisher={Geological Society of America}
+  publisher={Geological Society of America},
   doi ={https://doi.org/10.1130/GES02173.1}
 }
 @article{grosfils2011geologic,
@@ -86,7 +86,7 @@
   volume={8},
   number={85},
   pages={5300},
-  year={2023}
+  year={2023},
   doi ={10.21105/joss.05300}
 }
 @article{fractures,
@@ -97,7 +97,7 @@
   number={6},
   pages={603--624},
   year={2023},
-  publisher={Copernicus Publications G{\"o}ttingen, Germany}
+  publisher={Copernicus Publications G{\"o}ttingen, Germany},
   doi ={10.5194/se-14-603-2023}
 }
 @article{ernst,
@@ -108,7 +108,7 @@
   number={1-2},
   pages={1--58},
   year={1995},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi = {https://doi.org/10.1016/0012-8252(95)00017-5}
 }
 @manual{survey,
@@ -165,7 +165,7 @@
   number={1},
   pages={e2022JB026310},
   year={2023},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi ={10.1029/2022JB026310}
 }
 @article{scikit,
