### 2024-03-19 #6147 @aikubo
Thank you all for your patience @evetion @nialov @kthyng

I've finished addressing the reviews and have implemented major changes.

1. Added generic API with more data types
aikubo/LinkingLines/pull/27
With `geopandas` we can now read and write a much larger range of data types. I've also added test for this new functionality.

2. Added examples to the documentation
https://github.com/aikubo/LinkingLines/pull/28
- [Fracture data](https://linkinglines.readthedocs.io/en/latest/DemoFractures.html)
- [Venus Data](https://linkinglines.readthedocs.io/en/latest/DemoVenus.html)
These examples show that the package can be applied to a range of data types
- In addition, the documentation has been standardized so all docstrings are the same format
- Updated docs: https://linkinglines.readthedocs.io/en/latest/

3. Update Paper to focus on geosciences and misc corrections
- See updated paper here: https://github.com/aikubo/LinkingLines/actions/runs/8351474846

4. Addressed comments in issues opened by reviwers
https://github.com/aikubo/LinkingLines/issues/21
https://github.com/aikubo/LinkingLines/issues/21
https://github.com/aikubo/LinkingLines/issues/18
https://github.com/aikubo/LinkingLines/issues/17

5. Change package manager for ease of use and development
https://github.com/aikubo/LinkingLines/pull/29
- I've upgraded to using poetry for building and publishing the package.
- Tests and documentation now work fine with the local version
- I've also added a contributing guideline for using poetry

6. Add automatic tests and documentation build
https://github.com/aikubo/LinkingLines/pull/30
- Tests now run automatically for pull requests
- Documentation builds automatically when pushed to main
- This should also help for any future contributors

7. Authorship
- After reviewing the JOSS authorship criteria, I would like to clarify that the two listed coauthors made significant non-code contributions to this project including significant discussion of methods, design, and theory.

Please let me know any other comments or concerns.  I appreciate the comments which have helped make the package better, more generic, and more reproducible.
