### 2024-03-27 #6147 @evetion
Hi @aikubo, thanks for the update, there seem to be a lot of improvements! I've checked the review boxes for most now, except for the writing quality and references.

You're working on the JOSS branch now, might be good to merge to main/master soon, and let the bot here do it's work with generating a new version of the paper.

I see you are now the single author of the JOSS paper (is that correct?). Note I did not ask to remove the co-authors in my previous review, I merely asked for their roles, non-code authors are allowed.

Note that your code example in the README.md is out of date. Otherwise, the process of installing and testing went very smoothly with your recent changes/poetry. Kudos!

### Quality of writing

The paper has some odd sentences (l48-50), wording (statement of the field instead of state of the field) and characters (single `T` at the end of the summary). I would advise to carefully read it, use a spell/grammar checker, and ideally a colleague of yours reviews it.

### References
There are still several missing/wrong references in the linked pdf.
