### 2024-03-27 4f9020d
update paper and extensive grammar check for
https://github.com/openjournals/joss-reviews/issues/6147#issuecomment-2021865916
