+++ paper/paper.bib
@@ -36,9 +36,22 @@
  venue = {Austin, Texas},
  eventdate = {June 28 - July 3 2010}
 }
-
-
-@article{virtanen2020scipy,
+@article{survey,
+	title = {A survey of {Hough} {Transform}},
+	volume = {48},
+	issn = {00313203},
+	url = {http://dx.doi.org/10.1016/j.patcog.2014.08.027},
+	doi = {10.1016/j.patcog.2014.08.027},
+	abstract = {In 1962 Hough earned the patent for a method [1], popularly called Hough Transform (HT) that efficiently identifies lines in images. It is an important tool even after the golden jubilee year of existence, as evidenced by more than 2500 research papers dealing with its variants, generalizations, properties and applications in diverse fields. The current paper is a survey of HT and its variants, their limitations and the modifications made to overcome them, the implementation issues in software and hardware, and applications in various fields. Our survey, along with more than 200 references, will help the researchers and students to get a comprehensive view on HT and guide them in applying it properly to their problems of interest.},
+	number = {3},
+	journal = {Pattern Recognition},
+	author = {Mukhopadhyay, Priyanka and Chaudhuri, Bidyut B.},
+	year = {2015},
+	note = {Publisher: Elsevier},
+	keywords = {Digital Hough Transform (DHT), Generalized Hough Transform (GHT), Hough Transform (HT), Probabilistic Hough Transform (PHT), Randomized Hough Transform (RHT)},
+	pages = {993--1010},
+}
+@article{scipy,
   title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
   author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
   journal={Nature methods},
@@ -110,13 +123,6 @@
   year={1995},
   publisher={Elsevier},
   doi = {https://doi.org/10.1016/0012-8252(95)00017-5}
-}
-@manual{survey,
-  title        = {Bhukosh},
-  author       = {{GSI Bhukosh}},
-  year         = 2020,
-  note         = {[Computer software manual]},
-  url          = {http://bhukosh.gsi.gov.in}
 }
 @software{gpd,
   author       = {Kelsey Jordahl and
