+++ paper/paper.md
@@ -39,7 +39,7 @@ The `linkinglines` Python package addresses the need for quantitative and automa
 
 The primary needs that the `linkinglines` package fulfills include:
 
-1. **Dissected Line Extraction or Data Reduction**: In areas where land cover or data availabilty effects the complete mapping of linear features, this package can link together similarly oriented segments into lines in a quantitiative automated way. This is also an effective data reduction technique.
+1. **Dissected Line Extraction or Data Reduction**: In areas where land cover or data availability effects the complete mapping of linear features, this package can link together similarly oriented segments into lines in a quantitative automated way. This is also an effective data reduction technique.
 
 2. **Feature Extraction and Analysis**: The `linkinglines` package provides functions to extract and then compute essential metrics and statistics on extracted linear, radial or circumferential type features which are commonly seen in dike swarms or fracture networks.
 
@@ -51,25 +51,25 @@ The primary needs that the `linkinglines` package fulfills include:
 
 This package was originally developed to tackle the issue of mapped dike segments. Rugged terrain, vegetation cover, and a large area made it impossible to accurately map dikes. The length, density, and structure of the dike swarm affects how magma is transported and erupted. Scaling analysis indicated that for segments of widths of 10 m, dikes could be 10-10s of kilometers long however observed segments were two orders of magnitude lower [@morriss2020]. Additionally, the complex overlapping structure of the dike swarm was difficult to analyze. We designed `linkinglines` to extract not only lines from line segments but also help analyze the mesoscale structure of the dike swarm. Using the unique properties of the Hough Transform we can extract several unique mesoscale structures within a group of lines. 
 
-Issues with data coverage, incompleteness, and complexity occur and there are a wide range of Earth and Planetary science data and can also be applied including lineaments on Venus[@venus] and fracture networks[@fractures] of which we include as examples in our documentation. Clustering and machine learning mehtods are commonly applied geosciences within the subdisciplines of seismology, geochemistry, and planetary scinece [@li20f23machine]. There is currently available code to work with geospatial data clustering such as `geopandas`, `PySAL`, and others which perform the Hough Transform on images such as `scikit-image` or `OpenCV`, however this seems to be the first union of the two methods and specific to geoscience data applications [@gpd; @scikit; @opencv; @pysal]. There is a clear use case for `linkinglines`  to be applied to planetary science and the variety of linear features on Venus and Mars. For example, `linkinglines` can be used to quantitatively identify the centers of radial lineaments on Venus and relate the data back to volcanism and plate tectonics although work has been done to map and qualitatively characterize these features `linkinglines` offers unique insight and quantitative measure [@erst]. Related packages which are specific to geosciences such as `fractopo` which focuses on network analysis of fractures could be utilized in parallel with  `linkinglines` for additional analysis [@fractopo].
[…]
