### 2024-03-27 f9d6684
update citations for https://github.com/openjournals/joss-reviews/issues/6147
