### 2024-03-27 #6147 @aikubo
@evetion Thanks for pointing out the typos. I've gone over it again and hopefully fixed most of them and the citations for the paper.

I took the other authors off but after reviewing the criteria I added them back again and this version of the paper should reflect that.

I've also merged JOSS branch into master: https://github.com/aikubo/LinkingLines/pull/31
