### 2024-04-02 #6147 @evetion
> @evetion Thanks for pointing out the typos. I've gone over it again and hopefully fixed most of them and the citations for the paper.

Hi @aikubo, it seems my feedback about the writing quality and references are not yet fixed in the latest article proof. Maybe you missed it, or something is not merged yet?
