+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'LinkingLines: Using the Hough Transform to Cluster Line Segments and for
+title: 'LinkingLines: Using the Hough Transform to Cluster Line Segments and
 Mesoscale Feature Extraction'
 tags:
   - python
@@ -31,7 +31,7 @@ header-includes:
 
 # Summary
 
-Linear feature analysis plays a fundamental role in various scientific and geospatial applications, from detecting infrastructure networks to characterizing geological formations. In this paper, we introduce `linkinglines`, an open-source Python package tailored for the clustering, and feature extraction of linear structures in geospatial data. Our package leverages the Hough Transform, commonly used in image processing, performs clustering of line segments in the Hough Space then provides unique feature extraction methods and visualization. `linkinglines` empowers researchers, data scientists, and analysts across diverse domains to efficiently process, understand, and extract valuable insights from linear features, contributing to more informed decision-making and enhanced data-driven exploration. We have used `linkinglines` to map dike swarms with thousands of segments associated with Large Igneous Provinces in @kubo2023. T
+Linear feature analysis plays a fundamental role in various scientific and geospatial applications, from detecting infrastructure networks to characterizing geological formations. In this paper, we introduce `linkinglines`, an open-source Python package tailored for the clustering, and feature extraction of linear structures in geospatial data. Our package leverages the Hough Transform, commonly used in image processing, performs clustering of line segments in the Hough Space then provides unique feature extraction methods and visualization. `linkinglines` empowers researchers, data scientists, and analysts across diverse domains to efficiently process, understand, and extract valuable insights from linear features, contributing to more informed decision-making and enhanced data-driven exploration. We have used `linkinglines` to map dike swarms with thousands of segments associated with Large Igneous Provinces in @kubo2023. 
 
 # Statement of Need
 
@@ -39,17 +39,17 @@ The `linkinglines` Python package addresses the need for quantitative and automa
 
 The primary needs that the `linkinglines` package fulfills include:
 
-1. **Dissected Line Extraction or Data Reduction**: In areas where land cover or data availability effects the complete mapping of linear features, this package can link together similarly oriented segments into lines in a quantitative automated way. This is also an effective data reduction technique.
+1. **Dissected Line Extraction or Data Reduction**: In areas where land cover or data availability affects the complete mapping of linear features, this package can link together similarly oriented segments into lines in a quantitative automated way. This is also an data reduction technique.
 
 2. **Feature Extraction and Analysis**: The `linkinglines` package provides functions to extract and then compute essential metrics and statistics on extracted linear, radial or circumferential type features which are commonly seen in dike swarms or fracture networks.
 
 3. **Geospatial and Image Data Integration**: Geospatial data often involves complex relationships between various linear features. `linkinglines` integrates seamlessly with popular geospatial libraries like `pyproj` and `geopandas` to facilitate georeferenced data analysis and visualization.
 
-4. **Custom Plotting and Visualization**: Effective visualization is critical for data interpretation. The package offers custom plotting scripts, making it easier to visualize results and communicate findings effectively.
+4. **Custom Plotting and Visualization**: Effective visualization is critical for data interpretation. The package offers custom plotting scripts, making it easier to visualize results and communicate findings.
 
 ## Statement of the Field
 
[…]
