### 2024-04-02 #6147 @aikubo
> Hi @aikubo, it seems my feedback about the writing quality and references are not yet fixed in the latest article proof. Maybe you missed it, or something is not merged yet?

Yes, sorry about that, I didn't merge to master. It is there now.
