### 2024-05-10 #6147 @aikubo
Hello all just wanted to follow up on this @nialov  & @kthyng !

I'm doing my defense on 6/14 and I'd like to include the final published version in my dissertation. Not a big deal if not since it will get published any way but I wanted to check in.
