### 2024-05-13 aikubo/LinkingLines#33 @nialov
Issue with running `DemoLinkingLines.ipynb`
Trying to run `docs/DemoLinkingLines.ipynb`, I get the following error in cell number 2:

~~~
➜ poetry run ipython DemoLinkingLines.ipynb
Cell In[1], line 5
      1 # Load the example dataset
      2
      3 #load it using built in reading function which also does the preprocessing
      4 # the CSV must have a column called "WKT" and can have any other data
----> 5 dikeset=ll.readFile('../data/SpanishPeaks_3857.csv', preprocess=True)

File ~/projects/LinkingLines/src/linkinglines/PrePostProcess.py:84, in readFile(name, preprocess)
     82 # if preprocess is True, preprocess the data
     83 if preprocess:
---> 84     data = WKTtoArray(data)
     85     data = preProcess(data)
     87 return data

File ~/projects/LinkingLines/src/linkinglines/PrePostProcess.py:196, in WKTtoArray(df, plot)
    194 if not ("WKT" in df.columns ):
    195     if not ("geometry" in df.columns):
--> 196      raise ValueError("No geometry present")
    198 xstart=[]
    199 ystart=[]

ValueError: No geometry present
~~~

While making new data formats work, maybe you made a change that made the old data not work :sweat_smile:

There is also an error in the path to the file, it should be `'../data/SpanishPeaks_3857.csv'` rather than `'/../data/SpanishPeaks_3857.csv'` i.e. without the leading forward slash.
