### 2024-05-13 aikubo/LinkingLines#34 @nialov
Add short explanation inside `DemoFractures.ipynb`
I think `DemoFractures.ipynb` serves as a good example of using the software for other data than dykes but I think it looks like there are no radial patterns in the fracture data (not sure if I was expecting them when suggesting the data :laughing: ). Or that is how I interpret the results. I think you should add a few words to the notebook about there apparently not being any radial patterns according to the analysis in this data, which is a result in itself, so there are not any confusion in someone trying to interpret the results.
