### 2024-05-15 #6147 @nialov
I have updated the checklist above! If some checks are missing, see below text for the issue that I had.

# General checks

## Contribution and authorship

Thank you for the clarification!

## Substantial scholarly effort

Integration of `geopandas` has been done and the workflow has been generalised.

## Data sharing

Data is shared and used in the tests and documentation (notebooks).

# Functionality

## Functionality

I can now run the tests and execute the notebooks.

# Documentation

## Installation instructions

Development dependencies are documented.

## Example usage

Examples with fracture data and lineament data are included. However, few issues:

-  <https://github.com/aikubo/LinkingLines/issues/33>
-  <https://github.com/aikubo/LinkingLines/issues/34>

I especially like the Venus lineament example!

## Functionality documentation

`README.md` has been expanded!

## Automated tests

Tests are now automated and documented properly.

## Community guidelines

There now is a section, `## 4. Contributing Guidelines`, in the README.

There are some oddities however, such as this section:

~~~
You are now ready to utilize the power of Hough Line Transform, Agglomerative Clustering, and custom plotting in your data analysis projects. If you have any questions or need further assistance, please refer to the detailed documentation or contact our support team.
~~~

Please go through the documentation and remove all nonsensical (?) text such as this. As I see, you have not talked anything about a support team prior to this section? You should point users wanting support to e.g. an email address or to post issues on the *GitHub* issue page.

# Software paper

The *Example Code Usage* section has been removed.

The *Code Structure* section has been condensed. Inclusion of the two functions are now in the article seems fine to me.

However, the paper is still quite long (~1800 words) even with removals associated with this review. I would critically review the "Algorithm" section and see what parts have already been stated either in the previous article where this software was used or in the actual documentation of the software on *GitHub/ReadTheDocs*. References to the software packages you have used (e.g. `scipy`) are critical and you should keep those parts so they are cited properly! However, methodological details do not need to be included in this *JOSS* paper. Novel algorithmic details might be of interest and mentioned but many other details can be omitted, if they can be stated elsewhere (readme, actual documentation, etc.). See <https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain>. For example, the suggested maximum length is 1000 words. Also check that the equation numbers are correct in the paper. The numbers currently start from 3 rather than 1.

## Summary

I still find parts such as lines 12-15 to be overtly general. I would reword or remove to more specifically bring out what you can specifically do with this software.

## A statement of need

I find the third "need" to be a bit of a stretch (lines 33-36). I would just remove it and instead bring out integration with e.g. `geopandas` elsewhere, if wanted.

I find the sentence on lines 20 to 34 to be a bit nonsensical. Why does the growth in data volume require us to analyse and extract more linear features? I would reword to just bring out the need in the geological domain.

## State of the field

The heading is "Statement of the Field". Should it be "State"? Anyway, this section can be included in the "Statement of Need" section rather than have it separate. I would remove the title and integrate it into the previous section.

## Quality of writing

> Some parts are overly general and the paper could be more concise and focused on the specific issues that the software is trying to solve.

This is still an issue although most parts with this issue should now be highlighted in the previous sections of this post.

## References

I believe these are fine.
