### 2024-05-15 aikubo/linking-and-clustering-dikes#33 @nialov
Worked after [9d7a70d](https://github.com/aikubo/LinkingLines/commit/9d7a70da2c301516c144bcf2cb4496e44838e96a).

Closing, thanks!
