### 2024-05-15 aikubo/linking-and-clustering-dikes#34 @nialov
Sounds good after [ebc1d6a](https://github.com/aikubo/LinkingLines/commit/ebc1d6a33cf50ff8e92945fff0978570526675b2)

Closing!
