### 2024-05-17 091c964
clarify READme https://github.com/openjournals/joss-reviews/issues/6147
