### 2024-05-17 22ef156
reduce word count, streamline paper https://github.com/openjournals/joss-reviews/issues/6147
