### 2024-05-17 #6147 @aikubo
@nialov thank you for the comments

I've updated the examples so they should all run now:
https://github.com/aikubo/LinkingLines/issues/33

And I've worked streamlining the paper. I agree it is word.
I've gotten the WC down to ~1200 and removed the equation to simplify (they are in the documentation)

https://github.com/aikubo/LinkingLines/commit/22ef1560331dae628970b6bd7da452ff40fad402

I've also updated the readme:

https://github.com/aikubo/LinkingLines/commit/091c9648ec21b3a75e1997bd8f091e439c1a9a18
