### 2024-05-22 #6147 @hugoledoux
Thanks @evetion and @nialov , it seems we will manage before June 14.

I have proofread the paper and I fixed many issues and propose a few changes, see https://github.com/aikubo/LinkingLines/pull/37

Could you:
- [ ] merge the PR
- [ ] Fix the Equation 8 in Figure 2 (there are no equations in the paper)
- [ ] Fix the invalid DOIs, see above
- [ ] ensure that your affiliation is okay, in the PR I removed the departments so that both authors "1" have the same affiliation
