+++ paper/paper.bib
@@ -82,7 +82,7 @@
   pages={1082--1106},
   year={2020},
   publisher={Geological Society of America},
-  doi ={https://doi.org/10.1130/GES02173.1}
+  doi ={10.1130/GES02173.1}
 }
 @article{venus,
   title={Geologic map of the Ganiki Planitia quadrangle (V-14)},
@@ -122,7 +122,7 @@
   pages={1--58},
   year={1995},
   publisher={Elsevier},
-  doi = {https://doi.org/10.1016/0012-8252(95)00017-5}
+  doi = {10.1016/0012-8252(95)00017-5}
 }
 @software{gpd,
   author       = {Kelsey Jordahl and
@@ -199,7 +199,11 @@
     posted-at = {2008-01-15 19:21:54},
     priority = {4},
     title = {{The OpenCV Library}},
-    year = {2000}
+    volume={25},
+    number={11},
+    pages={120--123},
+    year={2000},
+    publisher={Miller Freeman Inc.}
 }
 @article{pysal,
     doi       = {10.21105/joss.02826},
