### 2024-05-23 65122b5
remove opencv bib entry
