+++ paper/paper.md
@@ -48,7 +48,7 @@ F
 
 This package was originally developed to tackle the issue of mapping dike segments. Rugged terrain, vegetation cover, and a large area made it impossible to accurately map dikes. The length, density, and structure of the dike swarm affects how magma is transported and erupted. Scaling analysis indicated that for segments of widths of 10 m, dikes could be tens to thousands of kilometers long; however observed segments were two orders of magnitude shorter [@morriss2020]. Additionally, the complex overlapping structure of the dike swarm was difficult to analyze. We designed `linkinglines` to extract not only lines from line segments but also help analyze the mesoscale structure of the dike swarm. Using the unique properties of the Hough Transform we can extract several unique mesoscale structures within a group of lines.
 
-Issues with data coverage, incompleteness, and complexity occur in a wide range of Earth and Planetary science data products. As an example we show in our documentation how this work can also be applied to lineaments on Venus [@venus] and fracture networks [@fractures]. Clustering and machine learning methods are commonly applied geosciences within the subdisciplines of seismology, geochemistry, and planetary science [@li2023machine]. There are currently available packages to work with geospatial data clustering, examples are `geopandas`, `PySAL`, and others which perform the Hough Transform on images such as `scikit-image` or `OpenCV`. However this seems to be the first union of the two methods and specific to geoscience data applications [@gpd; @scikit; @opencv; @pysal]. Related packages which are specific to geosciences, such as `fractopo` which focuses on network analysis of fractures, could be utilized in parallel with `linkinglines` for additional analysis [@fractopo].
+Issues with data coverage, incompleteness, and complexity occur in a wide range of Earth and Planetary science data products. As an example we show in our documentation how this work can also be applied to lineaments on Venus [@venus] and fracture networks [@fractures]. Clustering and machine learning methods are commonly applied geosciences within the subdisciplines of seismology, geochemistry, and planetary science [@li2023machine]. There are currently available packages to work with geospatial data clustering, examples are `geopandas`, `PySAL`, and others which perform the Hough Transform on images such as `scikit-image` or `OpenCV`. However this seems to be the first union of the two methods and specific to geoscience data applications [@gpd; @scikit; @pysal]. Related packages which are specific to geosciences, such as `fractopo` which focuses on network analysis of fractures, could be utilized in parallel with `linkinglines` for additional analysis [@fractopo].
 
 # Algorithm
 
