+++ paper/paper.bib
@@ -191,20 +191,6 @@
     url = {https://doi.org/10.7717/peerj.453},
     doi = {10.7717/peerj.453}
 }
-@article{opencv,
-    author = {Bradski, G.},
-    citeulike-article-id = {2236121},
-    journal = {Dr. Dobb's Journal of Software Tools},
-    keywords = {bibtex-import},
-    posted-at = {2008-01-15 19:21:54},
-    priority = {4},
-    title = {{The OpenCV Library}},
-    volume={25},
-    number={11},
-    pages={120--123},
-    year={2000},
-    publisher={Miller Freeman Inc.}
-}
 @article{pysal,
     doi       = {10.21105/joss.02826},
     url       = {https://doi.org/10.21105/joss.02826},
