### 2024-05-23 #6147 @aikubo
- [x]  merge the PR
- [x]  Fix the Equation 8 in Figure 2 (there are no equations in the paper)
- [ ]  Fix the invalid DOIs, see above
- [x]  ensure that your affiliation is okay, in the PR I removed the departments so that both authors "1" have the same affiliation

>> - No DOI given, and none found for title: The OpenCV Library

Oddly enough the openCV citation has no DOI that I can find I think because of it's age (2000) and it was originally published in a print magazine, it is the official one suggest by the OpenCV documentation. I think the easiest solution would be to just remove it.
