### 2024-05-24 #6147 @hugoledoux
> Oddly enough the openCV citation has no DOI that I can find I think because of it's age (2000) and it was originally published in a print magazine, it is the official one suggest by the OpenCV documentation. I think the easiest solution would be to just remove it.

👍
