### 2024-05-24 #6147 @hugoledoux
Okay, almost all is good. At this point could you:
- [ ] Fix the "F" that you wrongly added to the paper https://github.com/aikubo/LinkingLines/commit/8eabf59d774ff2b5bf0b4056ca1d509c2c0d4e7f#r142354645
