### 2024-05-26 #6147 @aikubo
> * [ X]  Fix the "F" that you wrongly added to the paper [aikubo/LinkingLines@8eabf59#r142354645](https://github.com/aikubo/LinkingLines/commit/8eabf59d774ff2b5bf0b4056ca1d509c2c0d4e7f#r142354645)

> * [x ]  Make a tagged release of your software, and list the version tag of the archived version here.
v2.1.2
tag here:  https://github.com/aikubo/LinkingLines/releases/tag/2.1.2

> * [x ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

https://zenodo.org/records/10887984

> * [X ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

> * [ X]  Please list the DOI of the archived version here.

10.5281/zenodo.10887984

Thank you for the excellent reviews from @nialov and @evetion ! Thank you all for bearing with me on this long review.
