### 2024-05-27 #6147 @hugoledoux
your zenodo has the wrong metadata, make sure to do this:

Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
