### 2024-05-28 #6147 @hugoledoux
hmmm, no the DOI is different, but okay I have the DOI.

However:

1. your name is different on the paper and on zenodo?! It has to be the same
2. order of the authors is also different.

You can fix both in the metadata without making a new release I reckon...
