### 2024-06-07 c72bcc2
update bib capitalization
