+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{hough,
-  title={Use of the Hough transformation to detect lines and curves in pictures},
+  title={Use of the {Hough} transformation to detect lines and curves in pictures},
   author={Duda, Richard O and Hart, Peter E},
   journal={Communications of the ACM},
   volume={15},
@@ -10,7 +10,7 @@
   doi = {10.1145/361237.361242}
 }
 @article{ballard1981generalizing,
-  title={Generalizing the Hough transform to detect arbitrary shapes},
+  title={Generalizing the {Hough} transform to detect arbitrary shapes},
   author={Ballard, Dana H.},
   journal={Pattern Recognition},
   volume={13},
@@ -22,7 +22,7 @@
 }
 @inproceedings{pandas,
  author = {McKinney, Wes},
- title = {Data Structures for Statistical Computing in Python},
+ title = {Data {Structures} for {Statistical} {Computing} in {Python}},
  pages = {56--61},
  bookpagination = {page},
  publisher = {SciPy},
@@ -52,7 +52,7 @@
 	pages = {993--1010},
 }
 @article{scipy,
-  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  title={SciPy 1.0: fundamental algorithms for scientific computing in {Python}},
   author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
   journal={Nature methods},
   volume={17},
@@ -74,7 +74,7 @@
   doi ={10.1029/2022GC010842}
 }
 @article{morriss2020,
-  title={The Chief Joseph dike swarm of the Columbia River flood basalts, and the legacy data set of William H. Taubeneck},
+  title={The {Chief} {Joseph} {Dike} {Swarm} of the {Columbia} {River} flood basalts, and the legacy data set of {William} {H.} {Taubeneck}},
   author={Morriss, Matthew C and Karlstrom, Leif and Nasholds, Morgan WM and Wolff, John A},
   journal={Geosphere},
   volume={16},
@@ -85,7 +85,7 @@
   doi ={10.1130/GES02173.1}
 }
 @article{venus,
-  title={Geologic map of the Ganiki Planitia quadrangle (V-14)},
+  title={Geologic map of the {Ganiki} {Planitia} quadrangle (V-14)},
   author={Grosfils, EB and Long, SM and Venechuk, EM and Hurwitz, DM and Richards, JW and Kastl, B and Drury, DE and Hardin, J},
   journal={Venus. United States Geological Survey, Scientific Investigations Map},
   volume={3121},
@@ -93,7 +93,7 @@
   doi ={10.3133/sim3121}
 }
 @article{fractopo,
-  title={fractopo: A Python package for fracture network analysis},
+  title={fractopo: {A} {Python} package for fracture network analysis},
   author={Ovaskainen, Nikolas},
   journal={Journal of Open Source Software},
   volume={8},
@@ -114,7 +114,7 @@
   doi ={10.5194/se-14-603-2023}
 }
 @article{ernst,
-  title={Giant radiating dyke swarms on Earth and Venus},
+  title={Giant radiating dyke swarms on {Earth} and {Venus}},
   author={Ernst, Richard E and Head, James W and Parfitt, Elisabeth and Grosfils, E and Wilson, L},
   journal={Earth-Science Reviews},
   volume={39},
@@ -175,7 +175,7 @@
   doi ={10.1029/2022JB026310}
 }
 @article{scikit,
-    title = {scikit-image: image processing in Python},
[…]
