### 2024-06-07 #6147 @aikubo
Fixed capitalization:
https://github.com/aikubo/LinkingLines/pull/38

Thanks for your patience, I've been prepping for my defense.
