### 2024-06-07 #6147 @kthyng
Congratulations on your new publication @aikubo, and good luck on your defense!! Many thanks to @hugoledoux and to reviewers @evetion and @nialov for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
